added documentation

+++ src/main/java/com/simonmittag/loggingutils/JvmMemoryUtil.java
@@ -5,24 +5,38 @@ package com.simonmittag.loggingutils;
  */
 public class JvmMemoryUtil {
 
+    /**
+     * Exec runtime env
+     */
     protected static Runtime runtime = Runtime.getRuntime();
 
+    /**
+     * Free MB Jvm memory
+     * @return Free MB Jvm memory
+     */
     public static float getFreeMb() {
-        return (float) (runtime.freeMemory() / 1024f / 1024f);
+        return (runtime.freeMemory() / 1024f / 1024f);
     }
 
+    /**
+     * Total MB Jvm memory
+     * @return Total MB Jvm memory
+     */
     public static float getTotalMb() {
-        return (float) (runtime.totalMemory() / 1024f / 1024f);
+        return (runtime.totalMemory() / 1024f / 1024f);
     }
 
+    /**
+     * Max MB Jvm memory
+     * @return Max MB Jvm memory
+     */
     public static float getMaxMb() {
-        return (float) (runtime.maxMemory() / 1024f / 1024f);
+        return (runtime.maxMemory() / 1024f / 1024f);
     }
 
     /**
      * Use this value object for printing and log statements.
-     *
-     * @return
+     * @return A wrapper obj with all Jvm stats
      */
     public static JvmMemory stats() {
         JvmMemory m = new JvmMemory();
